working partner tests
