+++ src/main/scala/krs/DecoratorPattern.scala
@@ -11,9 +11,7 @@ object DecoratorPattern {
     x
   }
 
-  def isEven(x: Int): (Nothing) => Boolean = {
-    logValue _ andThen evenCheck _ andThen logValue _
-  }
+  def isEven(x: Int): Boolean = logValue(evenCheck(logValue(x)))
 
   type CustomerId = Int
