+++ src/test/scala/krs/partner/InfrastructureModule.scala
@@ -1,5 +1,6 @@
 package krs.partner
 
+import com.twitter.util.Await
 import org.scalatest._
 
 class PartnerFileRepositorySpec extends FlatSpec with Matchers {
@@ -12,7 +13,7 @@ class PartnerFileRepositorySpec extends FlatSpec with Matchers {
 
   "loadOffers" should "have 7 items from CapitalOne" in {
     val repo = new PartnerFileRepository.Repository(fixtureData)
-    val offers = repo.loadOffers()
+    val offers = Await.result(repo.loadOffers())
     offers.length should be(7)
     offers(0).provider should be("CapitalOne")
   }
