+++ src/test/scala/krs/partner/PartnerDomain.scala
@@ -1,9 +1,7 @@
 package krs.partner
 
-import com.twitter.util.Await
+import com.twitter.util.{Await, Future}
 import org.scalatest._
-
-trait TestModule extends InfrastructureModule with DomainModule
 
 class PartnerSystemSpec extends FlatSpec with Matchers {
   import PartnerDomain._
@@ -19,38 +17,18 @@ class PartnerSystemSpec extends FlatSpec with Matchers {
     PersonalLoan("Offer08", Range(750, 770), 100.00, 12)
   )
 
-  "filterOffers" should "have 5 items for 550 score" in new TestModule {
-    val eligable = partnerApi.filterOffers(offers, 550)
+  "filterOffers" should "have 5 items for 550 score" in {
+    val eligable = PartnerSystem.filterOffers(offers, 550)
     eligable.length should be(5)
   }
 
-  "filterOffers" should "have 1 item for 750 score"  in new TestModule {
-    val eligable = partnerApi.filterOffers(offers, 750)
+  "filterOffers" should "have 1 item for 750 score" in {
+    val eligable = PartnerSystem.filterOffers(offers, 750)
     eligable.length should be(1)
   }
 
-  "getOffers" should "have 5 items for 550 score" in new TestModule {
-    val eligable = Await.result(partnerApi.getOffers(550))
+  "getOffers" should "have 5 items for 550 score" in {
+    val eligable = Await.result(PartnerSystem.getOffersFromRepo(Future.value(offers), 550))
     eligable.length should be(5)
-  }
-}
-
-class DecoratorPatternSpec extends FlatSpec with Matchers {
-  import DecoratorPattern._
-
-  "isEven" should "return true for an even number" in {
-    isEven(2) should be(true)
-  }
-
-  "isEven" should "return false for an odd number" in {
-    isEven(1) should be(false)
-  }
-
-  "getCustomer for id 1" should "return a user" in {
-    getCustomer(1).getOrElse(None) should not be(None)
-  }
-
-  "getCustomer for id 10" should "return not user" in {
-    getCustomer(10).getOrElse(None) should be(None)
   }
 }
